HYD-6172 Replace bunyan with custom logger
Some bunyan deps were blacklisted.

Replace bunyan with a custom logger.

The log format is rather simple, here is an example from the realtime service:

{
  "name":"realtime",
  "hostname":"lotus-30vm13",
  "pid":19872,
  "level":30,
  "sock": {
    "id": "/#z7g2SQ5zGkapWpjBAAAK"
  },
  "msg": "stream ended",
  "time": "2016-07-06T19:12:29.135Z",
  "v":0
}

Basically it's just a JSON object that is then searchable by bunyan-cli.

If we create a logger that uses the same format,
we can still use bunyan-cli to search / filter records.

  - Create module + tests.

Signed-off-by: Joe Grund <joe.grund@intel.com>